move hook to use-users
